+++ library/e2e/integration/standalone.e2e.ts
@@ -1,34 +1,28 @@
 describe('Standalone bundle', () => {
-  describe('.render()', () => {
-    before(() => {
-      cy.visit('e2e/standalone.html');
-    });
+  testSuite('With parser', 'e2e/sites/standalone.html');
+  testSuite('Without parser', 'e2e/sites/standalone-without-parser.html');
 
-    it('Container should exist', () => {
-      cy.get('#spec').should('exist');
-    });
+  function testSuite(testName: string, site: string) {
+    describe(testName, () => {
+      before(() => {
+        cy.visit(site);
+      });
 
-    it('Title of spec should be rendered', () => {
-      cy.contains('Streetlights API').should('exist');
-    });
+      it('Container should exist', () => {
+        cy.get('#spec').should('exist');
+      });
 
-    it('Servers of spec should be rendered', () => {
-      cy.contains('test.mosquitto.org:{port}').should('exist');
-    });
+      it('Title of spec should be rendered', () => {
+        cy.contains('Example AsyncAPI').should('exist');
+      });
 
-    it('Channels of spec should be rendered', () => {
-      cy.contains(
-        'smartylighting/streetlights/1/0/event/{streetlightId}/lighting/measured',
-      ).should('exist');
-      cy.contains(
-        'smartylighting/streetlights/1/0/action/{streetlightId}/turn/on',
-      ).should('exist');
-      cy.contains(
-        'smartylighting/streetlights/1/0/action/{streetlightId}/turn/off',
-      ).should('exist');
-      cy.contains(
-        'smartylighting/streetlights/1/0/action/{streetlightId}/dim',
-      ).should('exist');
+      it('Servers of spec should be rendered', () => {
+        cy.contains('example-server').should('exist');
+      });
+
+      it('Channels of spec should be rendered', () => {
+        cy.contains('example-channel').should('exist');
+      });
     });
-  });
+  }
 });
