+++ library/src/standalone-without-parser.ts
@@ -1,55 +1,10 @@
-import React from 'react';
-import {
-  hydrate as hydrateComponent,
-  render as renderComponent,
-} from 'react-dom';
-
 import AsyncApiComponent, {
   AsyncApiProps,
 } from './containers/AsyncApi/Standalone';
 
-function querySelector(selector: string): Element | DocumentFragment | null {
-  if (typeof document !== 'undefined') {
-    return document.querySelector(selector);
-  }
-  return null;
-}
-
-function render(
-  props: AsyncApiProps,
-  container?: Element | DocumentFragment | null,
-  callback?: () => void,
-) {
-  container = container || querySelector('asyncapi');
-  if (container === null) {
-    return;
-  }
-
-  renderComponent(
-    React.createElement(AsyncApiComponent, props),
-    container,
-    callback,
-  );
-}
-
-function hydrate(
-  props: AsyncApiProps,
-  container?: Element | DocumentFragment | null,
-  callback?: () => void,
-) {
-  container = container || querySelector('asyncapi');
-  if (container === null) {
-    return;
-  }
-
-  hydrateComponent(
-    React.createElement(AsyncApiComponent, props),
-    container,
-    callback,
-  );
-}
+import { createRender, createHydrate } from './standalone-codebase';
 
 export default {
-  render,
-  hydrate,
+  render: createRender<AsyncApiProps>(AsyncApiComponent),
+  hydrate: createHydrate<AsyncApiProps>(AsyncApiComponent),
 };
